% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{ancestors}
\alias{ancestors}
\title{Get Ancestors of a Node}
\usage{
ancestors(dag, node)
}
\arguments{
\item{dag}{A dag object}

\item{node}{Node name or index}
}
\value{
Character vector of ancestor node names
}
\description{
Get Ancestors of a Node
}
