% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_package}
\alias{check_package}
\title{Check if Package is Available}
\usage{
check_package(pkg, purpose = NULL)
}
\arguments{
\item{pkg}{Package name}

\item{purpose}{What the package is needed for}
}
\value{
Logical
}
\description{
Check if Package is Available
}
\keyword{internal}
