% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{children}
\alias{children}
\title{Get Children of a Node}
\usage{
children(dag, node)
}
\arguments{
\item{dag}{A dag object}

\item{node}{Node name or index}
}
\value{
Character vector of child node names
}
\description{
Get Children of a Node
}
