% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_test.R
\name{compare_dags}
\alias{compare_dags}
\title{Compare Two DAGs}
\usage{
compare_dags(
  data,
  dag1,
  dag2,
  B = 500,
  alpha = 0.05,
  parallel = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Numeric matrix with n observations and p variables}

\item{dag1}{First dag object}

\item{dag2}{Second dag object}

\item{B}{Number of bootstrap replicates}

\item{alpha}{Significance level}

\item{parallel}{Use parallel processing}

\item{verbose}{Show progress}
}
\value{
A list with test results for both DAGs and comparison
}
\description{
Tests which of two DAGs better fits the observed data.
}
\examples{
# Chain vs fork
A1 <- matrix(c(0, 1, 0, 0, 0, 1, 0, 0, 0), 3, 3, byrow = TRUE)
A2 <- matrix(c(0, 1, 1, 0, 0, 0, 0, 0, 0), 3, 3, byrow = TRUE)

g1 <- dag(A1, nodes = c("X1", "X2", "X3"))
g2 <- dag(A2, nodes = c("X1", "X2", "X3"))

set.seed(123)
X1 <- rnorm(200)
X2 <- 0.5 * X1 + rnorm(200, sd = 0.5)
X3 <- 0.5 * X2 + rnorm(200, sd = 0.5)
data <- cbind(X1, X2, X3)

compare_dags(data, g1, g2, B = 100)

}
