% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{count_orderings}
\alias{count_orderings}
\title{Compute Number of Topological Orderings}
\usage{
count_orderings(dag, exact = TRUE)
}
\arguments{
\item{dag}{A dag object}

\item{exact}{Logical; compute exact count (can be slow) or estimate}
}
\value{
Integer count of topological orderings
}
\description{
Computes or estimates the number of valid topological orderings for a DAG.
}
\details{
Computing the exact number of topological orderings is #P-complete,
so for large DAGs an estimate may be more practical.
}
