% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{dag}
\alias{dag}
\title{Create a DAG Object}
\usage{
dag(adjacency = NULL, edges = NULL, nodes = NULL)
}
\arguments{
\item{adjacency}{A square adjacency matrix where \code{A[i,j] = 1} indicates
an edge from node i to node j, or NULL if using \code{edges}.}

\item{edges}{A two-column matrix or data frame of edges, where each row
represents an edge from the first column to the second column.
Can be node names or indices.}

\item{nodes}{Character vector of node names. If NULL, uses column names of
adjacency matrix or generates names X1, X2, ...}
}
\value{
An S3 object of class "dag" with components:
\item{adjacency}{The adjacency matrix}
\item{nodes}{Character vector of node names}
\item{n}{Number of nodes}
}
\description{
Creates an S3 object representing a directed acyclic graph (DAG) from
an adjacency matrix or edge list.
}
\details{
The adjacency matrix convention is that \code{A[i,j] = 1} means there is
a directed edge from node i to node j (i.e., i is a parent of j).
The function validates that the graph is acyclic.
}
\examples{
# Create a chain DAG: X1 -> X2 -> X3
A <- matrix(c(0, 1, 0,
              0, 0, 1,
              0, 0, 0), nrow = 3, byrow = TRUE)
g <- dag(A, nodes = c("X1", "X2", "X3"))

# Create from edge list
edges <- data.frame(from = c("X1", "X2"), to = c("X2", "X3"))
g <- dag(edges = edges)

}
