% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{descendants}
\alias{descendants}
\title{Get Descendants of a Node}
\usage{
descendants(dag, node)
}
\arguments{
\item{dag}{A dag object}

\item{node}{Node name or index}
}
\value{
Character vector of descendant node names
}
\description{
Get Descendants of a Node
}
