% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.bootstrap_parallel}
\alias{.bootstrap_parallel}
\title{Parallel Bootstrap}
\usage{
.bootstrap_parallel(
  data,
  dag,
  ordering,
  cond_kdes,
  qn_result,
  h,
  h_params,
  B,
  n_obs,
  n_vars,
  verbose
)
}
\description{
Parallel Bootstrap
}
\keyword{internal}
