% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.bootstrap_sequential}
\alias{.bootstrap_sequential}
\title{Sequential Bootstrap}
\usage{
.bootstrap_sequential(
  data,
  dag,
  ordering,
  cond_kdes,
  qn_result,
  h,
  h_params,
  B,
  n_obs,
  n_vars,
  verbose
)
}
\description{
Sequential Bootstrap
}
\keyword{internal}
