% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qn_functional.R
\name{.compute_qn_quadrature}
\alias{.compute_qn_quadrature}
\title{Quadrature-based Q_n Computation}
\usage{
.compute_qn_quadrature(data, kde_list, bounds, n_vars, n_quad)
}
\description{
Quadrature-based Q_n Computation
}
\keyword{internal}
