% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{.enumerate_orderings}
\alias{.enumerate_orderings}
\title{Recursive Enumeration of Topological Orderings}
\usage{
.enumerate_orderings(adj, nodes, current, orderings_env, max_ord)
}
\description{
Recursive Enumeration of Topological Orderings
}
\keyword{internal}
