% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{.estimate_ordering_count}
\alias{.estimate_ordering_count}
\title{Estimate Ordering Count}
\usage{
.estimate_ordering_count(dag, n_samples = 1000)
}
\description{
Estimate Ordering Count
}
\keyword{internal}
