% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.fit_conditional_dists}
\alias{.fit_conditional_dists}
\title{Fit Conditional Distributions for Bootstrap}
\usage{
.fit_conditional_dists(data, dag, ordering)
}
\description{
Fit Conditional Distributions for Bootstrap
}
\keyword{internal}
