% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.gauss_legendre_compute}
\alias{.gauss_legendre_compute}
\title{Compute Gauss-Legendre Nodes and Weights}
\usage{
.gauss_legendre_compute(n)
}
\description{
Compute Gauss-Legendre Nodes and Weights
}
\keyword{internal}
