% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.gauss_legendre_lookup}
\alias{.gauss_legendre_lookup}
\title{Lookup Table for Small Gauss-Legendre Rules}
\usage{
.gauss_legendre_lookup(n)
}
\description{
Lookup Table for Small Gauss-Legendre Rules
}
\keyword{internal}
