% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.generate_bootstrap_sample}
\alias{.generate_bootstrap_sample}
\title{Generate One Bootstrap Sample}
\usage{
.generate_bootstrap_sample(n_obs, n_vars, cond_kdes)
}
\description{
Generate One Bootstrap Sample
}
\keyword{internal}
