% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_test.R
\name{.mi_test_all_orderings}
\alias{.mi_test_all_orderings}
\title{Test All Orderings}
\usage{
.mi_test_all_orderings(
  data,
  dag,
  test_functions,
  B,
  alpha,
  parallel,
  verbose,
  seed
)
}
\description{
Test All Orderings
}
\keyword{internal}
