% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{.rejection_sample_conditional}
\alias{.rejection_sample_conditional}
\title{Rejection Sampling from Conditional}
\usage{
.rejection_sample_conditional(kde, x_cond, n_samples)
}
\description{
Rejection Sampling from Conditional
}
\keyword{internal}
