% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.sample_conditional_kde}
\alias{.sample_conditional_kde}
\title{Sample from Conditional KDE}
\usage{
.sample_conditional_kde(kde, x_cond, n)
}
\description{
Sample from Conditional KDE
}
\keyword{internal}
