% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph.R
\name{extract_ancestral_subgraph}
\alias{extract_ancestral_subgraph}
\title{Extract Ancestral Subgraph}
\usage{
extract_ancestral_subgraph(dag, nodes)
}
\arguments{
\item{dag}{A dag object}

\item{nodes}{Character vector of target node names}
}
\value{
A dag object representing the ancestral subgraph
}
\description{
Extracts a subgraph containing specified nodes and all their ancestors.
}
\examples{
A <- matrix(c(0, 1, 0, 0,
              0, 0, 1, 0,
              0, 0, 0, 1,
              0, 0, 0, 0), 4, 4, byrow = TRUE)
g <- dag(A, nodes = c("X1", "X2", "X3", "X4"))

# Get ancestral subgraph of X4
anc_g <- extract_ancestral_subgraph(g, "X4")  # Includes all nodes

}
