% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{fisher_combine}
\alias{fisher_combine}
\title{Fisher's Combination Method}
\usage{
fisher_combine(p_values)
}
\arguments{
\item{p_values}{Vector of p-values}
}
\value{
A list with:
\item{combined_p}{Combined p-value}
\item{chi_sq}{Chi-squared statistic}
\item{df}{Degrees of freedom}
}
\description{
Combines p-values from multiple orderings using Fisher's method.
}
\details{
Fisher's method combines p-values using:
X = -2 * sum(log(p_i))
which follows a chi-squared distribution with 2k degrees of freedom
under the null hypothesis of all nulls being true.
}
