% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gauss_legendre}
\alias{gauss_legendre}
\title{Gauss-Legendre Quadrature Nodes and Weights}
\usage{
gauss_legendre(n)
}
\arguments{
\item{n}{Number of quadrature points}
}
\value{
List with nodes and weights
}
\description{
Gauss-Legendre Quadrature Nodes and Weights
}
\keyword{internal}
