% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering.R
\name{get_topological_orders}
\alias{get_topological_orders}
\title{Get All Topological Orderings}
\usage{
get_topological_orders(dag, max_orderings = 100)
}
\arguments{
\item{dag}{A dag object}

\item{max_orderings}{Maximum number of orderings to return (default 100)}
}
\value{
A list of character vectors, each representing a valid topological
ordering.
}
\description{
Enumerates all valid topological orderings of a DAG.
}
\details{
A topological ordering of a DAG is a linear ordering of vertices such that
for every directed edge (u, v), vertex u comes before v in the ordering.

The number of valid orderings can grow factorially with the number of nodes.
Use \code{max_orderings} to limit the enumeration.
}
\examples{
# Chain: only one ordering
A <- matrix(c(0, 1, 0, 0, 0, 1, 0, 0, 0), 3, 3, byrow = TRUE)
g <- dag(A, nodes = c("X1", "X2", "X3"))
get_topological_orders(g)

# Fork: multiple orderings
A <- matrix(c(0, 1, 1, 0, 0, 0, 0, 0, 0), 3, 3, byrow = TRUE)
g <- dag(A, nodes = c("X1", "X2", "X3"))
get_topological_orders(g)

}
