% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistic.R
\name{h_exp}
\alias{h_exp}
\title{Exponential Test Function}
\usage{
h_exp(x, t = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{t}{Parameter (default 1)}
}
\value{
exp(t * x)
}
\description{
Test function h(x) = exp(t * x).
}
\examples{
h_exp(c(-1, 0, 1, 2), t = 0.5)

}
