% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistic.R
\name{h_indicator}
\alias{h_indicator}
\title{Indicator Test Function}
\usage{
h_indicator(x, c = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{c}{Threshold (default 0)}
}
\value{
Numeric vector of 0s and 1s
}
\description{
Test function h(x) = I(x > c).
}
\examples{
h_indicator(c(-1, 0, 1, 2), c = 0.5)

}
