% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistic.R
\name{h_poly}
\alias{h_poly}
\title{Polynomial Test Function}
\usage{
h_poly(x, k = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{k}{Positive integer power (will be doubled, so h(x) = x^(2k))}
}
\value{
x^(2k)
}
\description{
Test function h(x) = x^(2k) for positive integer k.
}
\examples{
h_poly(c(-1, 0, 1, 2), k = 2)  # x^4

}
