% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistic.R
\name{h_squared}
\alias{h_squared}
\title{Squared Test Function}
\usage{
h_squared(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
x^2
}
\description{
The default test function h(x) = x^2.
}
\examples{
h_squared(c(-1, 0, 1, 2))

}
