% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{is_valid_ordering}
\alias{is_valid_ordering}
\title{Check if Ordering is Valid}
\usage{
is_valid_ordering(dag, ordering)
}
\arguments{
\item{dag}{A dag object}

\item{ordering}{Character vector of node names representing an ordering}
}
\value{
Logical indicating if the ordering is a valid topological order
}
\description{
Check if Ordering is Valid
}
