% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde_bivariate}
\alias{kde_bivariate}
\title{Bivariate Kernel Density Estimation}
\usage{
kde_bivariate(x, y, bandwidth = NULL, n_grid = 64)
}
\arguments{
\item{x}{Numeric vector for first variable}

\item{y}{Numeric vector for second variable}

\item{bandwidth}{Length-2 vector of bandwidths. If NULL, uses Silverman's rule
for each variable.}

\item{n_grid}{Number of grid points per dimension (default 64)}
}
\value{
An object of class "kde_bivariate" with components:
\item{x}{Grid for first variable}
\item{y}{Grid for second variable}
\item{z}{Matrix of density values}
\item{data}{Original data matrix}
\item{bandwidth}{Bandwidths used}
\item{n}{Sample size}
}
\description{
Estimates a bivariate density using product Gaussian kernel.
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100, sd = 0.5)
kde <- kde_bivariate(x, y)

}
