% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde_cond_eval}
\alias{kde_cond_eval}
\title{Evaluate Conditional Density}
\usage{
kde_cond_eval(kde, y_new, x_cond, truncate = 1e-10)
}
\arguments{
\item{kde}{A kde_conditional object}

\item{y_new}{Points at which to evaluate y density}

\item{x_cond}{Conditioning values for x}

\item{truncate}{Minimum density value (default 1e-10)}
}
\value{
Numeric vector of conditional density values
}
\description{
Evaluates conditional density p(y|x) at specified points.
}
