% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde_cond_sample}
\alias{kde_cond_sample}
\title{Sample from Conditional Distribution}
\usage{
kde_cond_sample(kde, x_cond, n_samples = 1)
}
\arguments{
\item{kde}{A kde_conditional object}

\item{x_cond}{Conditioning values}

\item{n_samples}{Number of samples per conditioning value}
}
\value{
Matrix of samples (rows = samples, cols = conditioning values)
}
\description{
Generates samples from an estimated conditional distribution p(y|x).
}
\details{
Uses rejection sampling with the marginal density of y as proposal.
}
