% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde_conditional}
\alias{kde_conditional}
\title{Conditional Density Estimation}
\usage{
kde_conditional(y, x, bandwidth_y = NULL, bandwidth_x = NULL, n_grid = 512)
}
\arguments{
\item{y}{Response variable (what we want the density of)}

\item{x}{Conditioning variable}

\item{bandwidth_y}{Bandwidth for y. If NULL, uses Silverman's rule.}

\item{bandwidth_x}{Bandwidth for x. If NULL, uses Silverman's rule.}

\item{n_grid}{Number of grid points for y (default 512)}
}
\value{
An object of class "kde_conditional" with components:
\item{y_grid}{Grid for y}
\item{data_y}{Original y data}
\item{data_x}{Original x data}
\item{bandwidth_y}{Bandwidth for y}
\item{bandwidth_x}{Bandwidth for x}
\item{n}{Sample size}
}
\description{
Estimates conditional density p(y|x) using kernel methods.
}
\details{
Uses the Nadaraya-Watson estimator approach where
p(y|x) = f(x,y) / f(x) is estimated via kernel smoothing.
}
\examples{
x <- rnorm(100)
y <- 2*x + rnorm(100, sd = 0.5)
kde <- kde_conditional(y, x)

}
