% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde_eval}
\alias{kde_eval}
\title{Evaluate KDE at Specific Points}
\usage{
kde_eval(kde, x_new, truncate = 1e-10)
}
\arguments{
\item{kde}{A kde_marginal object}

\item{x_new}{Points at which to evaluate density}

\item{truncate}{Minimum density value (default 1e-10)}
}
\value{
Numeric vector of density values
}
\description{
Evaluates a fitted KDE at arbitrary points.
}
\examples{
kde <- kde_marginal(rnorm(100))
kde_eval(kde, c(-1, 0, 1))

}
