% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_progress}
\alias{make_progress}
\title{Progress Bar Wrapper}
\usage{
make_progress(total, verbose = TRUE)
}
\arguments{
\item{total}{Total iterations}

\item{verbose}{Whether to show progress}
}
\value{
Progress bar object or NULL
}
\description{
Progress Bar Wrapper
}
\keyword{internal}
