% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{parents}
\alias{parents}
\title{Get Parents of a Node}
\usage{
parents(dag, node)
}
\arguments{
\item{dag}{A dag object}

\item{node}{Node name or index}
}
\value{
Character vector of parent node names
}
\description{
Get Parents of a Node
}
