% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{plot.dag}
\alias{plot.dag}
\title{Plot Method for DAG}
\usage{
\method{plot}{dag}(x, layout = "hierarchical", ...)
}
\arguments{
\item{x}{A dag object}

\item{layout}{Layout algorithm: "hierarchical" (default) or "circular"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot Method for DAG
}
