% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{print.dag}
\alias{print.dag}
\title{Print Method for DAG}
\usage{
\method{print}{dag}(x, ...)
}
\arguments{
\item{x}{A dag object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for DAG
}
