% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_test.R
\name{print.mi_test_all_orderings}
\alias{print.mi_test_all_orderings}
\title{Print Method for All-Orderings Test}
\usage{
\method{print}{mi_test_all_orderings}(x, ...)
}
\description{
Print Method for All-Orderings Test
}
