% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph.R
\name{problematic_subgraphs}
\alias{problematic_subgraphs}
\title{Find Problematic Subgraphs}
\usage{
problematic_subgraphs(subgraph_result)
}
\arguments{
\item{subgraph_result}{A subgraph_test_result object}
}
\value{
List of subgraph node sets that were rejected
}
\description{
Identifies subgraphs that show evidence against DAG compatibility.
}
