% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_parallel}
\alias{restore_parallel}
\title{Restore Parallel Backend}
\usage{
restore_parallel(old_plan)
}
\arguments{
\item{old_plan}{Previous plan from setup_parallel}
}
\description{
Restore Parallel Backend
}
\keyword{internal}
