% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_divide}
\alias{safe_divide}
\title{Safe Division}
\usage{
safe_divide(num, denom, default = 0)
}
\arguments{
\item{num}{Numerator}

\item{denom}{Denominator}

\item{default}{Value when denominator is near zero}
}
\value{
num/denom or default
}
\description{
Safe Division
}
\keyword{internal}
