% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scott_bandwidth}
\alias{scott_bandwidth}
\title{Scott's Rule for Bandwidth}
\usage{
scott_bandwidth(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Bandwidth value
}
\description{
Scott's Rule for Bandwidth
}
\keyword{internal}
