% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_parallel}
\alias{setup_parallel}
\title{Set Up Parallel Backend}
\usage{
setup_parallel(parallel = TRUE)
}
\arguments{
\item{parallel}{Logical or number of cores}
}
\value{
Previous plan (for restoration)
}
\description{
Set Up Parallel Backend
}
\keyword{internal}
