% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{silverman_bandwidth}
\alias{silverman_bandwidth}
\title{Silverman's Rule of Thumb for Bandwidth}
\usage{
silverman_bandwidth(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Bandwidth value
}
\description{
Silverman's Rule of Thumb for Bandwidth
}
\keyword{internal}
