% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tensor_quadrature}
\alias{tensor_quadrature}
\title{Create Tensor Product Quadrature Grid}
\usage{
tensor_quadrature(n_points, dims, bounds)
}
\arguments{
\item{n_points}{Number of points per dimension}

\item{dims}{Number of dimensions}

\item{bounds}{Matrix with 2 columns (lower, upper) and dims rows}
}
\value{
List with grid matrix and weights vector
}
\description{
Create Tensor Product Quadrature Grid
}
\keyword{internal}
