% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_class.R
\name{topological_orders}
\alias{topological_orders}
\title{Get All Topological Orderings}
\usage{
topological_orders(dag, max_orderings = 100)
}
\arguments{
\item{dag}{A dag object}

\item{max_orderings}{Maximum number of orderings to return (default 100).
Set to Inf for all orderings (may be slow for large DAGs).}
}
\value{
A list of character vectors, each representing a valid topological ordering
}
\description{
Enumerates all valid topological orderings of the DAG.
}
\details{
For a DAG with n nodes, there can be up to n! orderings. This function
uses a recursive algorithm with pruning to enumerate them efficiently.
}
\examples{
A <- matrix(c(0, 1, 1,
              0, 0, 1,
              0, 0, 0), nrow = 3, byrow = TRUE)
g <- dag(A, nodes = c("X1", "X2", "X3"))
topological_orders(g)

}
