% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_quadrature}
\alias{transform_quadrature}
\title{Transform Quadrature to Arbitrary Interval}
\usage{
transform_quadrature(gl, a, b)
}
\arguments{
\item{gl}{Gauss-Legendre rule (list with nodes and weights)}

\item{a}{Lower bound}

\item{b}{Upper bound}
}
\value{
Transformed rule
}
\description{
Transform Quadrature to Arbitrary Interval
}
\keyword{internal}
