% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data}
\alias{validate_data}
\title{Validate Data Matrix}
\usage{
validate_data(data, dag = NULL)
}
\arguments{
\item{data}{A numeric matrix or data frame}

\item{dag}{Optional dag object to check column matching}
}
\value{
The data as a matrix with column names
}
\description{
Validate Data Matrix
}
\keyword{internal}
